+++ CHANGELOG.md
@@ -1,8 +1,20 @@
+# v0.2.0 (2014-06-04)
+
+Now a resolved-AST is used. This has been enforced for the implementation of `@Delegate`
+
+- Add the optional named parameter `includePrivate` to `@ToString()` and `@EqualsAndHashCode`.
+- Implementation of `operator==(o)` has been changed to use `identical` and `runtimeType`.
+
+## Features
+
+- `@Delegate`: generate methods from a getter type.
+
 
 # v0.1.1 (2014-05-09)
 
 - Add the optional named parameter `callSuper` to `@ToString()` and `@EqualsAndHashCode`.
 - Add the optional named parameter `exclude` to `@ToString()` and `@EqualsAndHashCode`.
+
 
 # v0.1.0 (2014-05-08)
 
@@ -12,6 +24,7 @@ Initial release available for public testing with `@ToString()` and `@EqualsAndH
 
 - `@ToString()`: generate the implementation of `String toString()`.
 - `@EqualsAndHashCode()`: generate the implementation of `bool operator ==(o)` and `int get hashCode`.
+
 
 # Semantic Version Conventions